sync/x/sync/types/message_admin_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"sync/testutil/sample"
)

func TestMsgAdmin_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgAdmin
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgAdmin{
				Authority: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgAdmin{
				Authority: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
